move EnergySystem class to the groupings module
