+++ oemof/solph/__init__.py
@@ -2,9 +2,9 @@
 
 
 """
-from oemof.solph.network import (Sink, Source, LinearTransformer, Storage, Bus,
-                                 Flow, EnergySystem)
+from oemof.solph.network import (
+    Sink, Source, LinearTransformer, Storage, Bus, Flow)
 from oemof.solph.models import OperationalModel
-from oemof.solph.groupings import GROUPINGS
+from oemof.solph.groupings import (GROUPINGS, EnergySystem)
 from oemof.solph.options import (Investment, BinaryFlow, DiscreteFlow)
 from oemof.solph.plumbing import NodesFromCSV
